Add dummy user data and meeting slot functionality
This commit introduces a function for creating dummy data for StartupUser, PartnerUser, and AdminUser instances in the `investor_routes.py` file. The `user_models.py` file is updated to adjust the model structure of AdminUser and PartnerUser and to introduce a new ‘Slot’ model for scheduling meetings. The `operator_routes.py` and `operator_controller.py` files have been updated to add a new route and controller function for operators to add meeting slots. Furthermore, the way meetings are associated with users is adjusted to consider the newly introduced slot functionality.