Set secure attribute to true for user-session cookie
This commit updates the user-session cookie settings in both the startup_controller and auth_routes files. Previously, the 'secure' attribute was set to false, but it has now been changed to true to enhance the security of the application. This ensures that the user-session cookie is only sent over secure (HTTPS) connections.